requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
plotly>=5.10.0
statsmodels>=0.13.0
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
openai>=1.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2